// src/interface/types.tsx
// Navbar Container
export interface IFintro {
  readonly key: string;
  readonly text: string;
}

// Carousel Container
export interface IFpath {
  readonly key: string;
  readonly path: string;
  readonly text: string;
  readonly subText: string;
}

// Footer
export interface IFFooterData {
  readonly key: number;
  readonly data: string;
}